develop/vision/requirements.txt
numpy==1.23.4
opencv_python==4.6.0.66
pyserial==3.5
msgpack==1.1.0